Refactor: PRIZE 상수 prize라는 객체로 추상화
