+++ src/constants/view-messages.js
@@ -1,7 +1,7 @@
+import prize from '../domain/prize.js';
 import formatNumber from '../utils/FormatNumber.js';
 
 import { LOTTO_PRICE } from './lotto-constants.js';
-import { PRIZE } from './prize-constants.js';
 
 export const INPUT_MESSAGE = {
   PURCHASE_AMOUNT: '> 구입금액을 입력해 주세요. ',
@@ -15,9 +15,9 @@ export const OUTPUT_MESSAGE = {
   RESULT_HEADER: '\n당첨 통계\n--------------------',
 
   RESULT: (totalResult, rank) =>
-    `${PRIZE[rank].matchCount}개 일치${
-      PRIZE[rank].isBonus ? ', 보너스 볼 일치 ' : ' '
-    }(${formatNumber(PRIZE[rank].reward)}원) - ${totalResult[rank]}개`,
+    `${prize.getMatchCountByRank(rank)}개 일치${
+      prize.getIsBonusByRank(rank) ? ', 보너스 볼 일치 ' : ' '
+    }(${formatNumber(prize.getRewarByRank(rank))}원) - ${totalResult[rank]}개`,
 
   PROFIT: (profit) => `총 수익률은 ${formatNumber(profit)}%입니다.`,
 };
