+++ src/domain/LottoResultCalculator.js
@@ -1,6 +1,7 @@
-import { NO_MATCH_PLACE, PERCENTATION, PRIZE } from '../constants/prize-constants.js';
+import { NO_MATCH_PLACE, PERCENTATION } from '../constants/prize-constants.js';
 import excludeKeyFromObject from '../utils/excludeKeyFromObject.js';
 import roundToSecondDecimalPlace from '../utils/roundToSecondDecimalPlace.js';
+import prize from './prize.js';
 
 class LottoResultCalculator {
   #lottoList;
@@ -13,7 +14,7 @@ class LottoResultCalculator {
   }
 
   getTotalResult() {
-    const initialResult = this.#initializeResultObject();
+    const initialResult = prize.getInitiallResultObject();
     const totalResult = this.#lottoList.reduce((acc, lotto) => {
       const rank = lotto.getRank(this.#winningLotto);
       if (rank !== NO_MATCH_PLACE) {
@@ -25,24 +26,10 @@ class LottoResultCalculator {
     return excludeKeyFromObject({ object: totalResult, removeKey: NO_MATCH_PLACE });
   }
 
-  #initializeResultObject() {
-    const initialResult = Object.keys(PRIZE).reduce((acc, key) => {
-      acc[key] = 0;
-      return acc;
-    }, {});
-
-    initialResult[NO_MATCH_PLACE] = 0;
-
-    return initialResult;
-  }
-
   #getTotalReward() {
     const totalResult = this.getTotalResult();
-    const totalReward = Object.keys(totalResult).reduce((acc, cur) => {
-      const prizeReward = PRIZE[cur] ? PRIZE[cur].reward * totalResult[cur] : 0;
-      return acc + prizeReward;
-    }, 0);
-    return totalReward;
+
+    return prize.getTotalRewardByTotalResult(totalResult);
   }
 
   getProfit(purchaseAmount) {
