+++ src/domain/prize.js
@@ -1,0 +1,37 @@
+import { NO_MATCH_PLACE, PRIZE } from '../constants/prize-constants';
+
+const prize = {
+  getMatchCountByRank: (rank) => PRIZE[rank].matchCount,
+
+  getIsBonusByRank: (rank) => PRIZE[rank].isBonus,
+
+  getRewarByRank: (rank) => PRIZE[rank].reward,
+
+  getRankByMatchCountAndBonus: ({ numberMatchCount, isBonus }) => {
+    const entry = Object.entries(PRIZE).find(
+      ([_, value]) => value.matchCount === numberMatchCount && value.isBonus === isBonus,
+    );
+
+    return entry ? entry[0] : NO_MATCH_PLACE;
+  },
+
+  getInitiallResultObject: () => {
+    const initialResult = Object.keys(PRIZE).reduce((acc, key) => {
+      acc[key] = 0;
+      return acc;
+    }, {});
+
+    initialResult[NO_MATCH_PLACE] = 0;
+
+    return initialResult;
+  },
+
+  getTotalRewardByTotalResult: (totalResult) => {
+    return Object.keys(totalResult).reduce((acc, cur) => {
+      const prizeReward = PRIZE[cur] ? PRIZE[cur].reward * totalResult[cur] : 0;
+      return acc + prizeReward;
+    }, 0);
+  },
+};
+
+export default prize;
